fix(ci): keep binder helpers on approved paths
